\name{nmmso-package}
\alias{nmmso-package}
\alias{nmmso}
\docType{package}
\title{
\packageTitle{nmmso}
}
\description{
\packageDescription{nmmso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nmmso}
\packageIndices{nmmso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nmmso}

Maintainer: \packageMaintainer{nmmso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
